[HOTFIX] Auth 이슈 해결

+++ app/src/main/java/org/woowatechcamp/githubapplication/data/noti/NotiRepository.kt
@@ -9,5 +9,5 @@ class NotiRepository @Inject constructor(
 ) {
     suspend fun getNoti() = service.getNoti()
 
-    suspend fun markNoti(threadId : Int) = service.markNoti(threadId)
+    suspend fun markNoti(threadId : Long) = service.markNoti(threadId)
 }
